7.24.9 Немного рефакторинга - разбил редьюсер, разделив зоны ответственности: состояние приложения, загружаемые данные и пользователь.
