+++ src/reducer/name-space.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  DATA: `DATA`,
+  APP: `APP`,
+  USER: `USER`
+};
+
+export default NameSpace;
