+++ src/reducer/user/selectors.js
@@ -1,0 +1,6 @@
+import NameSpace from '../name-space';
+
+const NAME_SPACE = NameSpace.APP;
+
+export const getAuthorizationStatus = (state) =>
+  state[NAME_SPACE].authorizationStatus;
